No longer using the 'get_own_javascripts' method for refering the javascript modules exposed through a module, instead added them in the webpack (shim) configuration directly.
Instead - exposed a new method - 'get_javascript_entries' to be used
later by the new implementation of the webpack configuration. It will
utilize it to refer it as pluggable javascript module entry for any
particular module. It will allow us to reintroduce the pluggable
architecture.

Introduce a new browser tools javascript module, and will act as base
class for all tools, which needs to be registerd with the browser for
better initialization, and integration.