feat(oauth2): Add possibility to configure a custom Oauth2 username claim
There are cases where you don't want to use the email as username
when using Oauth2 authentication.
This feature provides the possibility to configure the Oauth2 claim
which should be used as username. The key in the config.py is called
'OAUTH2_USERNAME_CLAIM'. If you don't provide a custom key, the email
is used as username, like before. So it is completely backwards
compatible.