Remove dependency on simplejson package.
I don't think anybody needs compatiblity with Python 2.6 any more.

The only difference in API of the standard library json methods
and the ones from simplejson is attribute `encoding`, which has
been removed in the standard library (for good reasons). On the
other hand, the standard library now accepts both `str` and
`bytes`, so encoding is not necessary that much (and if it is
necessary, it should be done in the class specific encoder
anyway).

The only change in API is that
`web.pgadmin.utils.ajax.make_json_response()` function lost its
`encoding` parameter. It is never used in all pgadmin4 anyway
and given that `DataTypeJSONEncoder` produces pure Python `str`
anyway, I don't think it is a big loss.