Fix the filename for the zlib library.
Our build originally relied on a cmake/msbuild build of zlib,
however, this is not the recommended way of building zlib
(although it is the obvious one). We now build using nmake and
copy, which results in a minor change of the filename.

This is also consistent with what the PostgreSQL build system
expects.