Introduction of custom React Hook 'useSchemaState(...)'.
Current implementation of the schema data & state management using
the 'stateUtils' is very complicated, and added a lot of commplexity.

Introducing a React custom hook 'useSchemaState(...)', which will
allow us to manage the schema data & state management outside
the React components, and make the code more readable and testable.

Also - it will be able to reuse the schema data & state management
implementation outside the SchemaDialogView component.